Add quick settings panel and setup improvements
